Added connection SDK for Aylien Text Analysis API
The API seems to be working. We are limited to 1000 hits/day, though, this should still provide enough data. I have a few tweets to test how accurate it is. There is some discrepancy on how they rate some of the tweets and how I feel they should be rated; however, there is a confidence value that seems to be low on the analyses that I believe are wrong. We will just have to pick a threshold for this.